// workers/src/infrastructure/mq/types/mq.types.ts
export interface RTAssertQueueOptionsForMain extends amqp.Options.AssertQueue {
    deadLetterExchange: string; // এটি অবশ্যই থাকতে হবে
    deadLetterRoutingKey: string; // এটিও অবশ্যই থাকতে হবে
}

// রিট্রাই কিউ এর জন্য (REQUIRED to route back to Main Exchange)
export interface RTAssertQueueOptions extends amqp.Options.AssertQueue {
    messageTtl: number; // রিট্রাই কিউতে TTL থাকা বাধ্যতামূলক
    deadLetterExchange: string; // মেসেজ ড্রপ হলে মেইন এক্সচেঞ্জে ফেরত যাবে
    deadLetterRoutingKey: string; // মেইন কিউয়ের রাউটিং কি
}